src/dashboard/requirements.txt
Flask
Flask-Sockets
flask-cors
google-auth
google-cloud-pubsub
gunicorn
python-dateutil
python-dotenv
requests
